fix: always run limit fns asynchronously
This modifies the generator function returned from plimit to always
run functions passed to limit asynchronously.

This is done by modifying the enqueue method to always push the function
provided to the generator function onto the queue and then scheduling a
function to run asynchronously to dequeue and run that function.

I thought using using `Promise.resolve().then()` was a nice way to
schedule that dequeuing function asynchronously because it wasn't too
node-specific or browser-specific, but I'm happy to do something else
like preferring `process.nextTick` and then falling back to `setTimeout`
if `process.nextTick` isn't available.