Tighten up how and when tasks are started.
This code was full of explanations about why it is weird, but it
doesn't need to be weird. This appears to have been te result of
intersecting attempts to fix async store, consistent run behavior
on full or empty queues, and some performance improvements on
the resulting code.

This commit removes an extraneous microtick and the need to explain
why it exists by incrementing and decrementing activeCount on the
same side of a resolve() boundary.

Test added to prove that I haven't broken the system behavior by
removing this extra closure and await.